Optional id prefixing for chat commands
If two players with the name 'foo' are online:
ID=0 Name=foo (joined first)
ID=1 Name=foo (joined second)
Then the chat command:
/stats foo
always shows stats of foo that joined first (ID=0).
But now it is possible to do:
/stats 1:foo
And it will show stats of foo that joined second (ID=1).
/stats 1:bar
Will say that bar is not online no matter if the id is valid.